Initialize an LRMoE model based on Clustered Method of Moments

Initialize an LRMoE model based on Clustered Method of Moments

cmm_init(Y, X, n_comp, type, exact_Y = FALSE)

Arguments

Y

A N by d (exact_Y=T) or N by 4d (exact_Y=F) matrix of numerics, where N is sample size and d is the dimension of each obsevation. If the size is N by 4d, Each block of four columns should be organized as (tl, yl, yu, tu), representing the truncation lower bound, censoring lower bound, censoring upper bound and truncation upper bound.

X

X A N*P matrix of numerics, where P is the number of covariates. The first column of X should be 1, which is the intercept.

n_comp

Numeric, representing how many latent classes/groups to use.

type

A vector of strings of either "continuous" or "exact", representing whether each dimension of Y is continuous or discrete.

exact_Y

TRUE/FALSE: whether Y is observed exactly, or with censoring and/or truncation.

Value

A list where zero_y, mean_y_pos, var_y_pos, skewness_y_pos and kurtosis_y_pos represent the summary statistics of Yby dimension and by component. alpha_init and experts_init represents the parameter initializations. ll_init contains the loglikelihood of each experts fitted to Y by dimension and by component. ll_best suggests an initialization of expert functions based on the best loglikelihood.